server/requirements.txt
fastapi
uvicorn
pydantic
pydantic_core
pydantic-settings
psycopg2-binary
asyncpg
sqlalchemy
python-dotenv
python-slugify
alembic
email-validator
passlib[bcrypt]
bcrypt
python-jose
pytz
python-multipart
aiosmtplib
jinja2
watchfiles>=0.21.0
websockets>=12.0


langgraph
langgraph-sdk
langgraph-checkpoint-sqlite
langsmith
langchain-community
langchain-core
langchain-openai
notebook
tavily-python
wikipedia
trustcall
langgraph-cli